Update README with better dev version-server instructions.
