+++ README.md
@@ -4,10 +4,19 @@
 
 This project builds the [Micronaut Launch UI](https://launch.micronaut.io).
 
-## To generate the site run:
+## To build the site for CI run:
 
 ```bash
 ./gradlew build --console=plain
+```
+
+## For local development mimicing production launch site:
+
+- To bypass CORS restrictions, Launch the Local Version / Proxy Server
+
+```
+cd ./dev-proxy-server
+npm run start
 ```
 
 - Start the dev build of the site
@@ -16,3 +25,76 @@ This project builds the [Micronaut Launch UI](https://launch.micronaut.io).
 cd ./app/launch/
 npm run start:local
 ```
+
+## Running a Micronaut Starter API locally to dev against
+
+If you want to run against a version of the starter api (Micronaut Starter Api)[https://github.com/micronaut-projects/micronaut-starter]
+
+Outside of this project pull down that repo
+
+```bash
+$ git clone git@github.com:micronaut-projects/micronaut-starter.git
+$ cd micronaut-starter
+$ ./gradlew starter-web-netty:run
+```
+
+Then start up a version server to provide that instance
+
+```
+cd ./dev-proxy-server
+npm run start:starter
+```
+
+## Updating The Version feed
+
+You can find the version feed file at [./app/launch/public/mn-version-feed.json](./app/launch/public/mn-version-feed.json). It must be in valid JSON format, so beware of these gotyas!
+
+**Keys must be quoted**
+
+- GOOD
+
+  ```
+  {
+      "key": "RELEASE",
+      "baseUrl": "https://launch.micronaut.io",
+      "order": 0
+  }
+  ```
+
+- BAD
+
+  ```
+  {
+      key: "RELEASE",
+      baseUrl: "https://launch.micronaut.io",
+      order: 0
+  }
+  ```
+
+**No trailing comma**
+
+- GOOD
+
+  ```
+  {
+      "key": "RELEASE",
+      ...
+  },
+  {
+      "key": "NEXT",
+      ...
+  }
+  ```
+
+- BAD
+
+  ```
+  {
+      "key": "RELEASE",
+      ...
+  },
+  {
+      "key": "NEXT",
+      ...
+  },
+  ```
