+++ dev-proxy-server/run.command
@@ -1,5 +1,5 @@
 #!/bin/bash
 
 cd $(dirname $0)
-
+npm i
 npm run start
